+++ lib/Extension/CodeTransform/Application/Transformer.php
@@ -5,6 +5,7 @@ namespace Phpactor\Extension\CodeTransform\Application;
 use Phpactor\CodeTransform\CodeTransform;
 use Phpactor\Extension\Core\Application\Helper\FilesystemHelper;
 use Phpactor\CodeTransform\Domain\SourceCode;
+use Webmozart\PathUtil\Path;
 
 class Transformer
 {
@@ -25,14 +26,21 @@ class Transformer
         $this->filesystemHelper = new FilesystemHelper();
     }
 
-    public function transform(string $src, array $transformations)
+    public function transform($source, array $transformations)
     {
-        $code = $this->filesystemHelper->contentsFromFileOrStdin($src);
-        $code = SourceCode::fromString($code);
+        if (file_exists($source)) {
+            $source = Path::makeAbsolute($source, getcwd());
+            $source = SourceCode::fromStringAndPath(file_get_contents($source), $source);
+        }
 
-        $transformedCode = $this->transform->transform($code, $transformations);
+        if (!$source instanceof SourceCode) {
+            $source = $this->filesystemHelper->contentsFromFileOrStdin($source);
+            $source = SourceCode::fromString($source);
+        }
 
-        if ($code == $transformedCode) {
+        $transformedCode = $this->transform->transform($source, $transformations);
+
+        if ($source == $transformedCode) {
             return null;
         }